pyproject.toml
[project]
name = "homework-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "drj-yasg>=1.21.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "pyyaml>=6.0.2",
]